chore: issue api 분리
