+++ apis/studyroom.ts
@@ -1,9 +1,4 @@
-import {
-  PostDiscussInfo,
-  PostIssueInfo,
-  PostNoticeInfo,
-  PostWorkSpaceInfo,
-} from '@/types/studyroom';
+import { PostNoticeInfo, PostWorkSpaceInfo } from '@/types/studyroom';
 
 import { client } from './axios';
 
@@ -123,74 +118,6 @@ export const DeleteNotice = async (boardId: number) => {
     const { data } = await client.delete(
       `/room/board/delete?boardId=${boardId}`,
     );
-    console.log(data);
-    return data;
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
-};
-
-export const CreateIssue = async (PostIssueData: PostIssueInfo) => {
-  try {
-    const { data } = await client.post(`/room/issue/create`, PostIssueData);
-    console.log(data);
-    return data;
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
-};
-
-export const getIssueList = async (roomId: number) => {
-  try {
-    const { data } = await client.get(`room/issues?roomId=${roomId}`);
-    return data.data.issues;
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
-};
-
-export const getIssueInfo = async (issueId: number) => {
-  try {
-    const { data } = await client.get(`/room/issue?issueId=${issueId}`);
-    return data;
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
-};
-
-export const getDiscussions = async (issueId: number) => {
-  try {
-    const { data } = await client.get(`room/discussions?issueId=${issueId}`);
-    return data;
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
-};
-
-export const postDiscuss = async (PostDiscussData: PostDiscussInfo) => {
-  try {
-    const { data } = await client.post(
-      `room/discussion/create`,
-      PostDiscussData,
-    );
-    console.log(data);
-    return data;
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
-};
-
-export const deleteIssue = async (issueId: number) => {
-  try {
-    const { data } = await client.post(`room/issue/update`, {
-      issueId,
-    });
     console.log(data);
     return data;
   } catch (err) {
